Added handleChange method to ValidatedInput class component. Added ternary operators in render() method
