+++ react-validated-input-component/src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOMClient from 'react-dom/client';
-import ValidatedInput from './app-drawer';
+import ValidatedInput from './validated-input';
 
 const root = ReactDOMClient.createRoot(document.querySelector('#root'));
 root.render(<ValidatedInput />);
